Todas los metodos funcionando
