+++ app/controllers/application_controller.rb
@@ -1,15 +1,62 @@
 class ApplicationController < ActionController::Base
-  protect_from_forgery with: :exception
+	protect_from_forgery with: :exception
 
-  @@api_key = "o5bQnMbk@:BxrE"
+	@@api_key = "o5bQnMbk@:BxrE"
+	@@id_recepcion = "5cbd3ce444f67600049431c5"
+	@@id_despacho = "5cbd3ce444f67600049431c6"
+	@@id_pulmon = "5cbd3ce444f67600049431c9"
+	@@id_cocina = "5cbd3ce444f67600049431ca"
 
-  	# Funcion que hay que editar
 	def hashing(data, api_key)
 		hmac = OpenSSL::HMAC.digest(OpenSSL::Digest::Digest.new('sha1'), api_key.encode("ASCII"), data.encode("ASCII"))
 		signature = Base64.encode64(hmac).chomp
 		return signature
   	end
 
+  	# Funcionando bien
+  	def get_almacenes(api_key)
+		data = "GET"
+		hash_value = hashing(data, api_key)
+		almacenes = HTTParty.get('https://integracion-2019-dev.herokuapp.com/bodega/almacenes', 
+		  headers:{
+		    "Authorization": "INTEGRACION grupo4:#{hash_value}",
+		    "Content-Type": "application/json"
+		  })
+		puts "\nALMACENES\n"
+		puts JSON.pretty_generate(almacenes)
+		return almacenes
+	end
+
+	# Funcionando bien
+	def get_products_from_almacenes(api_key, almacenId, sku)
+		data = "GET#{almacenId}#{sku}"
+		hash_value = hashing(data, api_key)
+		products = HTTParty.get("https://integracion-2019-dev.herokuapp.com/bodega/stock?almacenId=#{almacenId}&sku=#{sku}",
+		  headers:{
+		    "Authorization": "INTEGRACION grupo4:#{hash_value}",
+		    "Content-Type": "application/json"
+		  })
+		puts "\nPRODUCTOS DE ALMACENES\n"
+		puts JSON.pretty_generate(products)
+		return products
+	end
+
+	# Funcionando bien
+	def get_products_from_almacenes_limit_primeros(api_key, almacenId, sku, limit)
+		data = "GET#{almacenId}#{sku}"
+		hash_value = hashing(data, api_key)
+		products = HTTParty.get("https://integracion-2019-dev.herokuapp.com/bodega/stock?almacenId=#{almacenId}&sku=#{sku}&limit=#{limit}",
+		  headers:{
+		    "Authorization": "INTEGRACION grupo4:#{hash_value}",
+		    "Content-Type": "application/json"
+		  })
+		puts "\nPRODUCTOS DE ALMACENES\n"
+		puts JSON.pretty_generate(products)
+		return products
+	end
+
+	# Funcionando bien
+	# Probado con la bodega del G14 5cbd3ce444f6760004943201
   	def mover_producto_entre_bodegas(api_key, productoId, almacenId, oc, precio)
 		data = "POST#{productoId}#{almacenId}"
 		hash_value = hashing(data, api_key)
@@ -17,8 +64,8 @@ class ApplicationController < ActionController::Base
 		  body:{
 		  	"productoId": productoId,
 		  	"almacenId": almacenId,
-		  	"oc": "null",
-		  	"precio": 0,
+		  	"oc": oc,
+		  	"precio": precio,
 		  }.to_json,
 		  headers:{
 		    "Authorization": "INTEGRACION grupo4:#{hash_value}",
@@ -28,5 +75,75 @@ class ApplicationController < ActionController::Base
 		puts JSON.pretty_generate(producto_movido)
 		return producto_movido
 	end
-  
+
+	# Funcionando bien
+	def mover_producto_entre_almacenes(producto_json, id_destino)
+		#productoId = producto_json["_id"]
+		productoId = producto_json
+		almacenId = id_destino
+
+		data = "POST#{productoId}#{almacenId}"
+		hash_value = hashing(data, @@api_key)
+		req = HTTParty.post("https://integracion-2019-dev.herokuapp.com/bodega/moveStock",
+		  body:{
+				"productoId": productoId,
+				"almacenId": almacenId,
+
+		  }.to_json,
+		  headers:{
+		    "Authorization": "INTEGRACION grupo4:#{hash_value}",
+		    "Content-Type": "application/json"
+		  })
+
+
+		puts "\nMOVER PRODUCTO ENTRE ALMACENES\n"
+		puts JSON.pretty_generate(req)
+		return req
+	end
+
+	# Funcionando bien
+	def obtener_skus_con_stock(api_key, almacenId)
+		data = "GET#{almacenId}"
+		hash_value = hashing(data, api_key)
+		skus = HTTParty.get("https://integracion-2019-dev.herokuapp.com/bodega/skusWithStock?almacenId=#{almacenId}",
+		  headers:{
+		    "Authorization": "INTEGRACION grupo4:#{hash_value}",
+		    "Content-Type": "application/json"
+		  })
+		puts "\nSKUS\n"
+		puts JSON.pretty_generate(skus)
+		return skus
+	end
+
+	# Funcionando bien
+	def fabricar_sin_pago(api_key, sku, cantidad)
+		data = "PUT#{sku}#{cantidad}"
+		hash_value = hashing(data, api_key)
+		products_produced = HTTParty.put("https://integracion-2019-dev.herokuapp.com/bodega/fabrica/fabricarSinPago",
+		  body:{
+		  	"sku": sku,
+		  	"cantidad": cantidad
+		  }.to_json,
+		  headers:{
+		    "Authorization": "INTEGRACION grupo4:#{hash_value}",
+		    "Content-Type": "application/json"
+		  })
+
+		puts "\nFABRICAR SIN PAGO\n"
+		puts JSON.pretty_generate(products_produced)
+		return products_produced
+	end
+
+	def fabricar_todo(api_key, lista_productos)
+		almacenes = (get_almacenes(api_key)).to_a
+		puts "..................."
+		for almacen in almacenes do
+			almacenId = almacen["_id"]
+			for producto in lista_productos
+				get_products_from_almacenes(api_key, almacenId, producto)
+			end
+		end
+		puts "..................."
+	end
 end
+
