+++ config/database.yml
@@ -23,7 +23,7 @@ default: &default
 
 development:
   <<: *default
-  database: SushiG4_development
+  database: sushig4_development
 
   # The specified database role being used to connect to postgres.
   # To create additional roles in postgres see `$ createuser --help`.
@@ -57,7 +57,7 @@ development:
 # Do not set this db to the same as development or production.
 test:
   <<: *default
-  database: SushiG4_test
+  database: sushig4_test
 
 # As with config/secrets.yml, you never want to store sensitive information,
 # like your database password, in your source code. If your source code is
@@ -80,6 +80,6 @@ test:
 #
 production:
   <<: *default
-  database: SushiG4_production
-  username: SushiG4
+  database: sushig4_production
+  username: sushig4
   password: <%= ENV['SUSHIG4_DATABASE_PASSWORD'] %>
